SSCS-8082 Adding Allowed Scenario 7
