+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/uc/scenarios/UcTemplateComponentId.java
@@ -5,7 +5,7 @@ public enum UcTemplateComponentId {
     CONFIRMED_OR_SET_ASIDE_PARAGRAPH,
     INSUFFICIENT_POINTS_PARAGRAPH,
     INSUFFICIENT_POINTS_PARAGRAPH_SCHEDULE_8_PARAGRAPH_4_APPLIED,
-    DOES_NOT_HAVE_LIMITED_CAPABILITY_FOR_WORK_PARAGRAPH,
+    DOES_HAVE_LIMITED_CAPABILITY_FOR_WORK_PARAGRAPH,
     DOES_NOT_HAVE_LIMITED_CAPABILITY_FOR_WORK_RELATED_ACTIVITY_PARAGRAPH,
     DOES_NOT_HAVE_LIMITED_CAPABILITY_FOR_SCHEDULE_9_PARAGRAPH_4_APPLIED_PARAGRAPH,
     DISEASE_OR_DISABLEMENT_PARAGRAPH,
