+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/uc/newscenarios/NewUcScenario7Test.java
@@ -1,17 +1,18 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.scenarios;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.newscenarios;
 
 import java.util.Arrays;
 import java.util.List;
 import org.junit.Assert;
 import org.junit.Test;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.UcTemplateContent;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.scenarios.UcScenario;
 import uk.gov.hmcts.reform.sscs.model.docassembly.Descriptor;
 import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody;
 
-public class UcScenario7Test {
+public class NewUcScenario7Test {
 
     @Test
-    public void testScenario7() {
+    public void testScenario1() {
 
         List<Descriptor> schedule6Descriptors =
             Arrays.asList(Descriptor.builder()
@@ -41,18 +42,11 @@ public class UcScenario7Test {
             + "\n"
             + "The decision made by the Secretary of State on 20/09/2020 is set aside.\n"
             + "\n"
-            + "Felix Sydney is to be treated as having limited capability for work.\n"
+            + "Felix Sydney has limited capability for work. The matter is now remitted to the Secretary of State to make a final decision upon entitlement to UC.\n"
             + "\n"
-            + "This is because insufficient points were scored to meet the threshold for the work capability assessment, but regulation 29 of the Employment and Support Allowance (ESA) Regulations 2008 applied.\n"
+            + "This is because insufficient points were scored to meet the threshold for the Work Capability Assessment, but Schedule 8, paragraph 4 of the UC Regulations 2013 applied.\n"
             + "\n"
-            + "In applying the work capability assessment 9 points were scored from the activities and descriptors in Schedule 2 of the ESA Regulations 2008 made up as follows:\n"
-            + "\n"
-            + "Mobilising Unaided\tc.1\t9\n"
-            + "\n"
-            + "\n"
-            + "Felix Sydney does not have limited capability for work-related activity because no descriptor from Schedule 3 applied.  Regulation 35 did not apply.\n"
-            + "\n"
-            + "The tribunal applied regulation 29 because there would be a substantial risk to the mental or physical health of any person if the appellant were found not to have limited capability for work.\n"
+            + "The tribunal applied Schedule 8, paragraph 4 because there would be a substantial risk to the mental or physical health of any person if the appellant were found not to have limited capability for work.\n"
             + "\n"
             + "My first reasons\n"
             + "\n"
@@ -63,14 +57,14 @@ public class UcScenario7Test {
             + "This has been an oral (face to face) hearing. Felix Sydney attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
             + "\n";
 
-        Assert.assertEquals(12, content.getComponents().size());
+        Assert.assertEquals(9, content.getComponents().size());
 
         Assert.assertEquals(expectedContent, content.toString());
 
     }
 
     @Test
-    public void testScenario7WhenNoSchedule6() {
+    public void testScenario1WhenNoSchedule6() {
 
         List<Descriptor> schedule6Descriptors =
             Arrays.asList();
@@ -83,7 +77,7 @@ public class UcScenario7Test {
                 .isAllowed(true)
                 .isSetAside(true)
                 .dateOfDecision("2020-09-20")
-                .ucNumberOfPoints(0)
+                .ucNumberOfPoints(9)
                 .pageNumber("A1")
                 .appellantName("Felix Sydney")
                 .reasonsForDecision(Arrays.asList("My first reasons", "My second reasons"))
@@ -97,15 +91,11 @@ public class UcScenario7Test {
             + "\n"
             + "The decision made by the Secretary of State on 20/09/2020 is set aside.\n"
             + "\n"
-            + "Felix Sydney is to be treated as having limited capability for work.\n"
+            + "Felix Sydney has limited capability for work. The matter is now remitted to the Secretary of State to make a final decision upon entitlement to UC.\n"
             + "\n"
-            + "This is because insufficient points were scored to meet the threshold for the work capability assessment, but regulation 29 of the Employment and Support Allowance (ESA) Regulations 2008 applied.\n"
+            + "This is because insufficient points were scored to meet the threshold for the Work Capability Assessment, but Schedule 8, paragraph 4 of the UC Regulations 2013 applied.\n"
             + "\n"
-            + "In applying the work capability assessment 0 points were scored from the activities and descriptors in Schedule 2 of the ESA Regulations 2008 made up as follows:\n"
-            + "\n"
-            + "Felix Sydney does not have limited capability for work-related activity because no descriptor from Schedule 3 applied.  Regulation 35 did not apply.\n"
-            + "\n"
-            + "The tribunal applied regulation 29 because there would be a substantial risk to the mental or physical health of any person if the appellant were found not to have limited capability for work.\n"
+            + "The tribunal applied Schedule 8, paragraph 4 because there would be a substantial risk to the mental or physical health of any person if the appellant were found not to have limited capability for work.\n"
             + "\n"
             + "My first reasons\n"
             + "\n"
@@ -116,7 +106,7 @@ public class UcScenario7Test {
             + "This has been an oral (face to face) hearing. Felix Sydney attended the hearing today and the Tribunal considered the appeal bundle to page A1. A Presenting Officer attended on behalf of the Respondent.\n"
             + "\n";
 
-        Assert.assertEquals(12, content.getComponents().size());
+        Assert.assertEquals(9, content.getComponents().size());
 
         Assert.assertEquals(expectedContent, content.toString());
